- Configure email server from the admin
For this the object com.salesmanager.core.modules.email.EmailConfig is serialized to JSON and saved in the database as a JSON string using MerchantConfiguration. Then from the MerchantConfigurationService you can retrieve the EmailCongig object so it can deserialize from JSON to EmailConfig. For a similar example look in ShippingServiceImpl which serializes and deserializes java beans

git-svn-id: http://svn.code.sf.net/p/shopizer/code/shopizer-portlet@2060 e80ee2c7-76f3-4005-aeaf-98396743f455